:lipstick: chore: large devices header created
