+++ src/components/Header/styles.js
@@ -1,4 +1,6 @@
 import styled from "styled-components";
+// components
+import { Col } from "react-bootstrap";
 
 export const HeaderWrapper = styled.header`
   & {
@@ -7,6 +9,18 @@ export const HeaderWrapper = styled.header`
 
     a {
       padding: initial;
+    }
+  }
+`;
+
+export const Nav = styled(Col)`
+  & {
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+
+    @media (min-width: 768px) {
+      display: block;
     }
   }
 `;
